# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SensorDataInput {
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Soil moisture raw value */
  soil_moisture: number;
  /** Light level raw value */
  light_level: number;
  /** Nitrogen NPK value */
  nitrogen: number;
  /** Phosphorus NPK value */
  phosphorus: number;
  /** Potassium NPK value */
  potassium: number;
  /** GPS latitude */
  latitude: number;
  /** GPS longitude */
  longitude: number;
}

export interface SensorData {
  id: number;
  temperature: number;
  humidity: number;
  soil_moisture: number;
  light_level: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  latitude: number;
  longitude: number;
  created_at: string;
}

export interface StatSummary {
  min: number | null;
  max: number | null;
  avg: number | null;
}

export interface SensorStats {
  temperature: StatSummary;
  humidity: StatSummary;
  soil_moisture: StatSummary;
  light_level: StatSummary;
  total_readings: number;
  last_reading_at: string | null;
}

export interface DeviceInfo {
  /** The ESP32 IP address or hostname */
  ip: string | null;
  /** Whether the device is currently configured */
  connected: boolean;
  /** Timestamp of last successful poll */
  last_polled_at: string | null;
  /** Last poll error message if any */
  last_error: string | null;
}

export interface DeviceUpdateInput {
  /** The ESP32 IP address or hostname (e.g. 192.168.1.100) */
  ip: string;
}

export interface ErrorResponse {
  error: string;
}

export type GetSensorHistoryParams = {
  /**
   * @minimum 1
   * @maximum 100
   */
  limit?: number;
};