# lib/api-zod/src/generated/types/deviceInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceInfo {
  /** The ESP32 IP address or hostname */
  ip: string | null;
  /** Whether the device is currently configured */
  connected: boolean;
  /** Timestamp of last successful poll */
  last_polled_at: Date | null;
  /** Last poll error message if any */
  last_error: string | null;
}