# lib/api-zod/src/generated/types/deviceUpdateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface DeviceUpdateInput {
  /** The ESP32 IP address or hostname (e.g. 192.168.1.100) */
  ip: string;
}