# lib/api-zod/src/generated/types/sensorData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface SensorData {
  id: number;
  temperature: number;
  humidity: number;
  soil_moisture: number;
  light_level: number;
  nitrogen: number;
  phosphorus: number;
  potassium: number;
  latitude: number;
  longitude: number;
  created_at: Date;
}