# lib/api-zod/src/generated/types/sensorDataInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface SensorDataInput {
  /** Temperature in Celsius */
  temperature: number;
  /** Humidity percentage */
  humidity: number;
  /** Soil moisture raw value */
  soil_moisture: number;
  /** Light level raw value */
  light_level: number;
  /** Nitrogen NPK value */
  nitrogen: number;
  /** Phosphorus NPK value */
  phosphorus: number;
  /** Potassium NPK value */
  potassium: number;
  /** GPS latitude */
  latitude: number;
  /** GPS longitude */
  longitude: number;
}