# lib/api-zod/src/generated/types/sensorStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */
import type { StatSummary } from "./statSummary";

export interface SensorStats {
  temperature: StatSummary;
  humidity: StatSummary;
  soil_moisture: StatSummary;
  light_level: StatSummary;
  total_readings: number;
  last_reading_at: Date | null;
}