# lib/api-zod/src/generated/types/statSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * RRF Smart Farming System API
 * OpenAPI spec version: 0.1.0
 */

export interface StatSummary {
  min: number | null;
  max: number | null;
  avg: number | null;
}